Fixed docking motor code

+++ src/FCCode/controllers/master/apply_uplink_helpers.cpp
@@ -5,18 +5,18 @@
 
 void Master::apply_uplink_constants(const Comms::Uplink& uplink) {
   std::array<unsigned int, 5> constant_ids {
-     uplink.constant_0_id,
-     uplink.constant_1_id,
-     uplink.constant_2_id,
-     uplink.constant_3_id,
-     uplink.constant_4_id,
+     (unsigned int) uplink.constant_0_id,
+     (unsigned int) uplink.constant_1_id,
+     (unsigned int) uplink.constant_2_id,
+     (unsigned int) uplink.constant_3_id,
+     (unsigned int) uplink.constant_4_id,
   };
   std::array<unsigned int, 5> constant_vals {
-     uplink.constant_0_val,
-     uplink.constant_1_val,
-     uplink.constant_2_val,
-     uplink.constant_3_val,
-     uplink.constant_4_val,
+     (unsigned int) uplink.constant_0_val,
+     (unsigned int) uplink.constant_1_val,
+     (unsigned int) uplink.constant_2_val,
+     (unsigned int) uplink.constant_3_val,
+     (unsigned int) uplink.constant_4_val,
   };
   for(int i = 0; i < constant_ids.size(); i++) {
     unsigned int const_id = constant_ids[i];
